// src/store/mutation-types.ts
export enum MutationTypes {
  SET_INFO = "SET_INFO",
  SET_REQ_STATUS = "SET_REQ_STATUS",
  SET_SEARCH = "SET_SEARCH",
  SET_STATUS = "SET_STATUS",
  SET_CHARACTERS = "SET_CHARACTERS",
  SET_CHARACTERS_LOADING = "SET_CHARACTERS_LOADING",
  SET_CHARACTER = "SET_CHARACTER",
  SET_CHARACTER_BY_SEARCH = "SET_CHARACTER_BY_SEARCH",
  SET_CHARACTER_LOADING = "SET_CHARACTER_LOADING",
  SET_EPISODES_DETAILS = "SET_EPISODES_DETAILS",
  SET_EPISODES_DETAILS_DONE = "SET_EPISODES_DETAILS_DONE",
  SET_EPISODE_CHARACTERS = "SET_EPISODE_CHARACTERS",
  SET_EPISODE_SELECTED = "SET_EPISODE_SELECTED",
  SET_SHOW_EPISODE_CHARACTERS = "SET_SHOW_EPISODE_CHARACTERS",
}